+++ src/logger.js
@@ -1,11 +1,17 @@
 import { deps } from "./deps.js";
 
+const secretValues = new Set();
+
 class BrightConsoleHandler extends deps.logging.BaseHandler {
   /**
    * @override
    */
   format(logRecord) {
     let msg = super.format(logRecord);
+
+    for (const secret of secretValues) {
+      msg = msg.replaceAll(secret, "*****");
+    }
 
     switch (logRecord.level) {
       case deps.logging.LogLevels.INFO:
@@ -49,5 +55,15 @@ deps.logging.log.setup({
 });
 
 const logger = deps.logging.log.getLogger();
+
+export const registerSecrets = (secrets) => {
+  if (secrets && typeof secrets === "object") {
+    Object.values(secrets).forEach((value) => {
+      if (value && typeof value === "string" && value.length > 0) {
+        secretValues.add(value);
+      }
+    });
+  }
+};
 
 export default logger;