+++ test/test.js
@@ -1,39 +1,56 @@
-import { assertSnapshot } from "https://deno.land/std@0.185.0/testing/snapshot.ts";
-
-const run = async (cmd) => {
-  const command = new Deno.Command("sh", {
-    args: ["-c", cmd],
-    stdout: "piped",
-    stderr: "piped",
-  });
-
-  const { code, stdout, stderr } = await command.output();
-
-  const removeAnsi = (input) => {
-    // deno-lint-ignore no-control-regex
-    const ansiEscapeSequences = /\u001b\[[0-9;]*[a-zA-Z]/g;
-    return input.replace(ansiEscapeSequences, "");
-  };
-
-  const removeDuration = (input) => {
-    const durationRegEx = /obtained in \d+ms/g;
-    return input.replace(durationRegEx, "obtained in *ms");
-  };
-
-  let output = new TextDecoder().decode(stdout);
-  let outputError = new TextDecoder().decode(stderr);
-
-  output = removeDuration(removeAnsi(output));
-  outputError = removeAnsi(outputError);
-
-  return {
-    code,
-    output,
-    outputError,
-  };
-};
+import { createTestRunner } from "./test-utils.js";
 
 Deno.test("Given API", async (t) => {
+  const apiProcess = await startTestApi();
+  const test = createTestRunner(t);
+
+  await test("jsonr -e test get.http", "test/requests/api2");
+  await test("jsonr -e test post.http", "test/requests/api2");
+  await test("jsonr -e test delete.http", "test/requests/api2");
+  await test("jsonr -e test put.http", "test/requests/api2");
+  await test("jsonr -e test exception.http", "test/requests/api2");
+  await test("jsonr -e test get-auth-401.http -s 401", "test/requests/api1");
+  await test("jsonr -e test get-auth.http", "test/requests/api1");
+  await test('jsonr -m PUT -b \'{"name":"test"}\' localhost:3000/sample');
+  await test("jsonr http://localhost:3000/sample");
+  await test("jsonr -e test put.http -s 303", "test/requests/api2");
+  await test("jsonr -e test put.http -s 200", "test/requests/api2");
+  await test("jsonr -e test get.http -t test", "test/requests/api2");
+  await test(
+    "jsonr -e test -h 'X-SampleHeader: abc' -v get-auth.http",
+    "test/requests/api1",
+  );
+  await test(
+    "jsonr -e test -h 'X-SampleHeader: abc' -v get.http",
+    "test/requests/api2",
+  );
+  await test("jsonr -e test get.http -t sample-get", "test/requests/api2");
+  await test("jsonr http://localhost:3000/redirect");
+  await test("jsonr http://localhost:3000/redirect -f");
+  await test("jsonr -e test --js post-js.http", "test/requests/api2");
+  await test(
+    "jsonr --js -m POST -b '{ name: \"test\", count: 8 }' localhost:3000/sample",
+  );
+  await test("jsonr help");
+  await test("jsonr -e nonExistentEnv get.http", "test/requests/api2");
+  await test("jsonr config", "test/requests/api1");
+
+  await test("jsonr run --init http://localhost:3000/sample");
+  try {
+    await Deno.remove("jsonr-script.js");
+  } catch {
+    // Ignore if file doesn't exist
+  }
+
+  await test("jsonr run jsonr-script.js", "test/requests");
+  await test("jsonr run jsonr-script.js -v", "test/requests");
+  await test("jsonr run jsonr-script.js -e test -v", "test/requests/api1");
+
+  apiProcess.kill();
+  await apiProcess.output();
+});
+
+const startTestApi = async () => {
   const apiCommand = new Deno.Command("deno", {
     args: "run -A test-api.js".split(" "),
     stdout: "piped",
@@ -55,73 +72,5 @@ Deno.test("Given API", async (t) => {
     await new Promise((resolve) => setTimeout(resolve, 100));
   } while (true);
 
-  const test = async (jsonrCommand) => {
-    jsonrCommand = jsonrCommand.replace("jsonr", "deno run -A ../main.js");
-    await t.step(jsonrCommand, async () => {
-      const { code, output, outputError } = await run(jsonrCommand);
-
-      await assertSnapshot(t, { code, output, outputError });
-    });
-  };
-
-  const sdkTest = async (jsonrCommand) => {
-    jsonrCommand = jsonrCommand.replace("jsonr", "deno run -A ../main.js");
-    await t.step(jsonrCommand, async () => {
-      const initResult = await run(jsonrCommand);
-
-      const scriptResult = await run(
-        "deno run -A ../main.js run jsonr-script.js",
-      );
-
-      const normalizeScriptOutput = (output) => {
-        return output
-          .replace(/Elapsed: \d+ ms/g, "Elapsed: * ms")
-          .replace(/Header date: [^\n]+/g, "Header date: *");
-      };
-
-      await assertSnapshot(t, {
-        initCode: initResult.code,
-        initOutput: initResult.output,
-        initOutputError: initResult.outputError,
-        scriptCode: scriptResult.code,
-        scriptOutput: normalizeScriptOutput(scriptResult.output),
-        scriptOutputError: scriptResult.outputError,
-      });
-
-      try {
-        await Deno.remove("jsonr-script.js");
-      } catch {
-        // Ignore if file doesn't exist
-      }
-    });
-  };
-
-  await test("jsonr requests/get.http");
-  await test("jsonr requests/post.http");
-  await test("jsonr requests/delete.http");
-  await test("jsonr requests/put.http");
-  await test("jsonr requests/exception.http");
-  await test("jsonr -s 401 requests/auth-401.http");
-  await test("jsonr -e requests/environments/test.json requests/auth.http");
-  await test(
-    'jsonr -m PUT -b \'{"name":"test"}\' localhost:3000/sample',
-  );
-  await test("jsonr http://localhost:3000/sample");
-  await test("jsonr requests/put.http -s 303");
-  await test("jsonr requests/put.http -s 200");
-  await test("jsonr requests/get.http -t test");
-  await test("jsonr requests/get.http -t sample-get");
-  await test("jsonr http://localhost:3000/redirect");
-  await test("jsonr http://localhost:3000/redirect -f");
-  await test("jsonr --js requests/post-js.http");
-  await test(
-    "jsonr --js -m POST -b '{ name: \"test\", count: 8 }' localhost:3000/sample",
-  );
-  await test("jsonr help");
-
-  await sdkTest("jsonr init requests/get.http");
-  await sdkTest("jsonr init http://localhost:3000/sample");
-
-  apiProcess.kill();
-  await apiProcess.output();
-});
+  return apiProcess;
+};